virtual_screening/requirements.txt
# SciPy Stack.  Order is important!
matplotlib
numpy
pandas

# OE Stack
openeye-floe==0.2.157
#openeye-aquifer

# Others
requests
